move useHover to Tooltip.tsx
